Moved `component/Registry` into `store` Renamed module `component` into `components`; No longer exposing `Registry` type.
