+++ chartlets.js/packages/lib/src/components/Children.test.tsx
@@ -1,7 +1,7 @@
 import type { FC } from "react";
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { render, screen } from "@testing-library/react";
-import { registry } from "@/component/Registry";
+import { registry } from "@/components/registry";
 import { type ComponentProps } from "./Component";
 import { Children } from "./Children";
